# Add Cython accelerated fANOVA package.