--- a/.github/workflows/github-release.yml
+++ b/.github/workflows/github-release.yml
@@ -26,7 +26,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip setuptools
-          pip install --progress-bar off wheel twine
+          python -m pip install --progress-bar off wheel twine cython
       - run: python setup.py sdist bdist_wheel
       - run: twine check dist/*