--- a/optuna_dashboard/_fast_fanova/__init__.py
+++ b/optuna_dashboard/_fast_fanova/__init__.py
@@ -1,0 +1,1 @@
+from .evaluator import FanovaImportanceEvaluator  # NOQA