--- a/python_tests/test_fast_fanova/test_fanova.py
+++ b/python_tests/test_fast_fanova/test_fanova.py
@@ -8,7 +8,7 @@ import pytest
 try:
     from optuna_dashboard._fast_fanova import FanovaImportanceEvaluator
     from optuna.distributions import FloatDistribution
-except:
+except ImportError:
     FanovaImportanceEvaluator = None  # type: ignore
     FloatDistribution = None  # type: ignore