--- a/setup.cfg
+++ b/setup.cfg
@@ -43,6 +43,8 @@ optuna_dashboard =
     public/*
     img/*
     index.html
+    _fast_fanova/*.pyx
+    _fast_fanova/*.c
 
 [options.packages.find]
 exclude =