--- a/setup.cfg
+++ b/setup.cfg
@@ -27,6 +27,8 @@ classifiers =
 python_requires = >=3.6
 include_package_data = False
 packages = find:
+setup_requires =
+    numpy
 install_requires =
     optuna>=2.4
     packaging