--- a/setup.cfg
+++ b/setup.cfg
@@ -29,6 +29,7 @@ include_package_data = False
 packages = find:
 setup_requires =
     numpy
+    scikit-learn
 install_requires =
     optuna>=2.4
     packaging