--- a/setup.cfg
+++ b/setup.cfg
@@ -53,7 +53,7 @@ exclude =
 [flake8]
 max-line-length = 99
 statistics = True
-exclude = venv,build,.tox
+exclude = venv,build,.tox,.eggs,tmp
 
 [mypy]
 ignore_missing_imports = True