--- a/setup.py
+++ b/setup.py
@@ -1,26 +1,32 @@
 import os
 
-from setuptools import setup, Extension
+from setuptools import Extension
+from setuptools import setup
 from setuptools.command.build_ext import build_ext
 
 
-class LazyImportBuildExtCmd(build_ext):
-    def run(self):
-        import numpy
-        self.include_dirs.append(numpy.get_include())
-
 try:
     from Cython.Build import cythonize
+
     ext = ".pyx"
 except ImportError:
     cythonize = None
     ext = ".c"
 
+
+class LazyImportBuildExtCmd(build_ext):
+    def run(self):
+        import numpy
+
+        self.include_dirs.append(numpy.get_include())
+        super().run()
+
+
 ext_modules = [
     Extension(
         "optuna_dashboard._fast_fanova.tree",
         sources=[os.path.join("optuna_dashboard", "_fast_fanova", "tree" + ext)],
-        language='c'
+        language="c",
     )
 ]
 
@@ -31,4 +37,5 @@ if cythonize is not None:
 if __name__ == "__main__":
     setup(
         ext_modules=ext_modules,
+        cmdclass={"build_ext": LazyImportBuildExtCmd},
     )