@c-bata 2022-05-06
```python
import os
import time
import optuna
from optuna.importance import MeanDecreaseImpurityImportanceEvaluator, FanovaImportanceEvaluator
from optuna_dashboard._fast_fanova.evaluator import FastFanovaImportanceEvaluator


def objective(trial):
    x = trial.suggest_float("x", -100, 100)
    y = trial.suggest_float("y", -1, 1)
    z = trial.suggest_int("z", -10, 10)
    foo = trial.suggest_categorical("foo", ["foo", "bar"])
    return x**2 + y + z + len(foo)


def create_study(filepath):
    db_url = f"sqlite:///{filepath}"
    if os.path.exists(filepath):
        return optuna.load_study(storage=db_url, study_name="foo")

    study = optuna.create_study(storage=db_url, study_name="foo")
    study.optimize(objective, n_trials=1000)
    return study


if __name__ == '__main__':
    study = create_study("profile-fanova.db")
    print("n_trials", study._storage.get_n_trials(study._study_id))

    for n_trees in [1, 4, 8, 16, 32]:
        for (name, evaluator_cls) in [
            ("MeanDecreaseImpurity", MeanDecreaseImpurityImportanceEvaluator),
            ("Fanova (Optuna)     ", FanovaImportanceEvaluator),
            ("Fanova (FastFanova) ", FastFanovaImportanceEvaluator),
        ]:
            start = time.time()
            importances = optuna.importance.get_param_importances(
                study, evaluator=evaluator_cls(n_trees=n_trees, seed=0)
            )
            elapsed = time.time() - start
            print(f"{name}\tn_trees={n_trees}\t{elapsed:.3f}s\t{dict(importances)}")
```

```
$ python benchmark.py
n_trials 1000
Fanova (Optuna)     	n_trees=1	1.136s	{'x': 0.9998646755620669, 'y': 7.60978540848553e-05, 'z': 5.627026236357675e-05, 'foo': 2.956321484570648e-06}
Fanova (FastFanova) 	n_trees=1	0.193s	{'x': 0.9998646755620669, 'y': 7.609785408485662e-05, 'z': 5.627026236357677e-05, 'foo': 2.956321484570518e-06}
Fanova (Optuna)     	n_trees=4	3.972s	{'x': 0.999739753121187, 'y': 0.00019338699040713752, 'foo': 4.376565551336532e-05, 'z': 2.30942328925442e-05}
Fanova (FastFanova) 	n_trees=4	0.571s	{'x': 0.999739753121187, 'y': 0.00019338699040713544, 'foo': 4.3765655513366114e-05, 'z': 2.309423289254405e-05}
Fanova (Optuna)     	n_trees=8	8.009s	{'x': 0.9997525059167741, 'y': 0.0001801998425495004, 'foo': 4.5589378723504794e-05, 'z': 2.1704861952803143e-05}
Fanova (FastFanova) 	n_trees=8	1.050s	{'x': 0.9997525059167741, 'y': 0.00018019984254949933, 'foo': 4.558937872350498e-05, 'z': 2.1704861952803093e-05}
Fanova (Optuna)     	n_trees=16	15.671s	{'x': 0.9992140398988923, 'y': 0.0006487184521432268, 'z': 0.00010231547349331292, 'foo': 3.492617547105853e-05}
Fanova (FastFanova) 	n_trees=16	1.987s	{'x': 0.9992140398988923, 'y': 0.0006487184521432272, 'z': 0.00010231547349331384, 'foo': 3.492617547105858e-05}
Fanova (Optuna)     	n_trees=32	31.067s	{'x': 0.9990320336465125, 'y': 0.0006233616696319889, 'z': 0.0002820024134077058, 'foo': 6.26022704477642e-05}
Fanova (FastFanova) 	n_trees=32	3.911s	{'x': 0.9990320336465125, 'y': 0.0006233616696319885, 'z': 0.0002820024134077061, 'foo': 6.260227044776459e-05}
Fanova (Optuna)     	n_trees=64	61.678s	{'x': 0.9990239210567721, 'y': 0.0006636977385551106, 'z': 0.00019779759689672645, 'foo': 0.00011458360777611123}
Fanova (FastFanova) 	n_trees=64	7.702s	{'x': 0.9990239210567721, 'y': 0.0006636977385551106, 'z': 0.0001977975968967265, 'foo': 0.00011458360777611146}
```