@c-bata 2022-05-09
I will publish this package in separated repository to make this available for non optuna-dashboard users.